+++ public/app/features/scopes/testUtils/mocks.ts
@@ -1,30 +1,6 @@
-import { screen } from '@testing-library/react';
-import { KBarProvider } from 'kbar';
-import { render } from 'test/test-utils';
-
 import { Scope, ScopeDashboardBinding, ScopeNode } from '@grafana/data';
-import {
-  AdHocFiltersVariable,
-  behaviors,
-  GroupByVariable,
-  sceneGraph,
-  SceneGridItem,
-  SceneGridLayout,
-  SceneQueryRunner,
-  SceneTimeRange,
-  SceneVariableSet,
-  VizPanel,
-} from '@grafana/scenes';
-import { AppChrome } from 'app/core/components/AppChrome/AppChrome';
-import { DashboardControls } from 'app/features/dashboard-scene/scene//DashboardControls';
-import { DashboardScene } from 'app/features/dashboard-scene/scene/DashboardScene';
 
-import { ScopesFacade } from './ScopesFacadeScene';
-import { scopesDashboardsScene, scopesSelectorScene } from './instance';
-import { getInitialDashboardsState } from './internal/ScopesDashboardsScene';
-import { initialSelectorState } from './internal/ScopesSelectorScene';
-import * as api from './internal/api';
-import { DASHBOARDS_OPENED_KEY } from './internal/const';
+import * as api from '../internal/api';
 
 export const mocksScopes: Scope[] = [
   {
@@ -434,160 +410,3 @@ export const getMock = jest
 
     return {};
   });
-
-const selectors = {
-  tree: {
-    search: 'scopes-tree-search',
-    headline: 'scopes-tree-headline',
-    select: (nodeId: string, type: 'result' | 'persisted') => `scopes-tree-${type}-${nodeId}-checkbox`,
-    radio: (nodeId: string, type: 'result' | 'persisted') => `scopes-tree-${type}-${nodeId}-radio`,
-    expand: (nodeId: string, type: 'result' | 'persisted') => `scopes-tree-${type}-${nodeId}-expand`,
-    title: (nodeId: string, type: 'result' | 'persisted') => `scopes-tree-${type}-${nodeId}-title`,
-  },
-  selector: {
-    input: 'scopes-selector-input',
-    container: 'scopes-selector-container',
-    loading: 'scopes-selector-loading',
-    apply: 'scopes-selector-apply',
-    cancel: 'scopes-selector-cancel',
-  },
-  dashboards: {
-    expand: 'scopes-dashboards-expand',
-    container: 'scopes-dashboards-container',
-    search: 'scopes-dashboards-search',
-    loading: 'scopes-dashboards-loading',
-    dashboard: (uid: string) => `scopes-dashboards-${uid}`,
-    notFoundNoScopes: 'scopes-dashboards-notFoundNoScopes',
-    notFoundForScope: 'scopes-dashboards-notFoundForScope',
-    notFoundForFilter: 'scopes-dashboards-notFoundForFilter',
-    notFoundForFilterClear: 'scopes-dashboards-notFoundForFilter-clear',
-  },
-};
-
-export const getSelectorInput = () => screen.getByTestId<HTMLInputElement>(selectors.selector.input);
-export const querySelectorApply = () => screen.queryByTestId(selectors.selector.apply);
-export const getSelectorApply = () => screen.getByTestId(selectors.selector.apply);
-export const getSelectorCancel = () => screen.getByTestId(selectors.selector.cancel);
-
-export const getDashboardsExpand = () => screen.getByTestId(selectors.dashboards.expand);
-export const queryDashboardsContainer = () => screen.queryByTestId(selectors.dashboards.container);
-export const queryDashboardsSearch = () => screen.queryByTestId(selectors.dashboards.search);
-export const getDashboardsSearch = () => screen.getByTestId<HTMLInputElement>(selectors.dashboards.search);
-export const queryAllDashboard = (uid: string) => screen.queryAllByTestId(selectors.dashboards.dashboard(uid));
-export const queryDashboard = (uid: string) => screen.queryByTestId(selectors.dashboards.dashboard(uid));
-export const getDashboard = (uid: string) => screen.getByTestId(selectors.dashboards.dashboard(uid));
-export const getNotFoundNoScopes = () => screen.getByTestId(selectors.dashboards.notFoundNoScopes);
-export const getNotFoundForScope = () => screen.getByTestId(selectors.dashboards.notFoundForScope);
-export const getNotFoundForFilter = () => screen.getByTestId(selectors.dashboards.notFoundForFilter);
-export const getNotFoundForFilterClear = () => screen.getByTestId(selectors.dashboards.notFoundForFilterClear);
-
-export const getTreeSearch = () => screen.getByTestId<HTMLInputElement>(selectors.tree.search);
-export const getTreeHeadline = () => screen.getByTestId(selectors.tree.headline);
-export const getResultApplicationsExpand = () => screen.getByTestId(selectors.tree.expand('applications', 'result'));
-export const queryResultApplicationsGrafanaTitle = () =>
-  screen.queryByTestId(selectors.tree.title('applications-grafana', 'result'));
-export const getResultApplicationsGrafanaTitle = () =>
-  screen.getByTestId(selectors.tree.title('applications-grafana', 'result'));
-export const getResultApplicationsGrafanaSelect = () =>
-  screen.getByTestId(selectors.tree.select('applications-grafana', 'result'));
-export const queryPersistedApplicationsGrafanaTitle = () =>
-  screen.queryByTestId(selectors.tree.title('applications-grafana', 'persisted'));
-export const queryResultApplicationsMimirTitle = () =>
-  screen.queryByTestId(selectors.tree.title('applications-mimir', 'result'));
-export const getResultApplicationsMimirTitle = () =>
-  screen.getByTestId(selectors.tree.title('applications-mimir', 'result'));
-export const getResultApplicationsMimirSelect = () =>
-  screen.getByTestId(selectors.tree.select('applications-mimir', 'result'));
-export const queryPersistedApplicationsMimirTitle = () =>
-  screen.queryByTestId(selectors.tree.title('applications-mimir', 'persisted'));
-export const getPersistedApplicationsMimirTitle = () =>
-  screen.getByTestId(selectors.tree.title('applications-mimir', 'persisted'));
-export const getPersistedApplicationsMimirSelect = () =>
-  screen.getByTestId(selectors.tree.select('applications-mimir', 'persisted'));
-export const queryResultApplicationsCloudTitle = () =>
-  screen.queryByTestId(selectors.tree.title('applications-cloud', 'result'));
-export const getResultApplicationsCloudSelect = () =>
-  screen.getByTestId(selectors.tree.select('applications-cloud', 'result'));
-export const getResultApplicationsCloudExpand = () =>
-  screen.getByTestId(selectors.tree.expand('applications-cloud', 'result'));
-export const getResultApplicationsCloudDevSelect = () =>
-  screen.getByTestId(selectors.tree.select('applications-cloud-dev', 'result'));
-export const getResultApplicationsCloudOpsSelect = () =>
-  screen.getByTestId(selectors.tree.select('applications-cloud-ops', 'result'));
-
-export const getResultCloudSelect = () => screen.getByTestId(selectors.tree.select('cloud', 'result'));
-export const getResultCloudExpand = () => screen.getByTestId(selectors.tree.expand('cloud', 'result'));
-export const getResultCloudDevRadio = () =>
-  screen.getByTestId<HTMLInputElement>(selectors.tree.radio('cloud-dev', 'result'));
-export const getResultCloudOpsRadio = () =>
-  screen.getByTestId<HTMLInputElement>(selectors.tree.radio('cloud-ops', 'result'));
-export const getResultCloudProdRadio = () =>
-  screen.getByTestId<HTMLInputElement>(selectors.tree.radio('cloud-prod', 'result'));
-
-export function buildTestScene(overrides: Partial<DashboardScene> = {}) {
-  return new DashboardScene({
-    title: 'hello',
-    uid: 'dash-1',
-    description: 'hello description',
-    tags: ['tag1', 'tag2'],
-    editable: true,
-    $timeRange: new SceneTimeRange({
-      timeZone: 'browser',
-    }),
-    controls: new DashboardControls({}),
-    $behaviors: [
-      new behaviors.CursorSync({}),
-      new ScopesFacade({
-        handler: (facade) => sceneGraph.getTimeRange(facade).onRefresh(),
-      }),
-    ],
-    $variables: new SceneVariableSet({
-      variables: [
-        new AdHocFiltersVariable({
-          name: 'adhoc',
-          datasource: { uid: 'my-ds-uid' },
-        }),
-        new GroupByVariable({
-          name: 'groupby',
-          datasource: { uid: 'my-ds-uid' },
-        }),
-      ],
-    }),
-    body: new SceneGridLayout({
-      children: [
-        new SceneGridItem({
-          key: 'griditem-1',
-          x: 0,
-          y: 0,
-          width: 300,
-          height: 300,
-          body: new VizPanel({
-            title: 'Panel A',
-            key: 'panel-1',
-            pluginId: 'table',
-            $data: new SceneQueryRunner({ key: 'data-query-runner', queries: [{ refId: 'A' }] }),
-          }),
-        }),
-      ],
-    }),
-    ...overrides,
-  });
-}
-
-export function renderDashboard(dashboardScene: DashboardScene) {
-  return render(
-    <KBarProvider>
-      <AppChrome>
-        <dashboardScene.Component model={dashboardScene} />
-      </AppChrome>
-    </KBarProvider>
-  );
-}
-
-export function resetScenes() {
-  scopesSelectorScene?.setState(initialSelectorState);
-
-  localStorage.removeItem(DASHBOARDS_OPENED_KEY);
-
-  scopesDashboardsScene?.setState(getInitialDashboardsState());
-}
